refactor(skills): simplify management mutations
