+++ backend/tests/unit/onyx/skills/test_skill_patch_unset_sentinel.py
@@ -8,32 +8,30 @@ from onyx.server.features.skill.models import SkillPatchRequest
 
 
 def test_omitted_fields_do_not_count_as_updates() -> None:
-    req = SkillPatchRequest(is_public=True)
+    req = SkillPatchRequest(public_permission=SkillSharePermission.VIEWER)
 
-    assert req.model_fields_set == {"is_public"}
+    assert req.model_fields_set == {"public_permission"}
     assert req.has_db_field_update is True
     assert req.has_details_update is False
 
 
 def test_all_fields_sent() -> None:
     req = SkillPatchRequest(
-        is_public=True,
         public_permission=SkillSharePermission.EDITOR,
         enabled=False,
     )
 
-    assert req.model_fields_set == {"is_public", "public_permission", "enabled"}
-    assert req.is_public is True
+    assert req.model_fields_set == {"public_permission", "enabled"}
     assert req.public_permission == SkillSharePermission.EDITOR
     assert req.enabled is False
 
 
-def test_false_values_count_as_updates() -> None:
-    req = SkillPatchRequest(is_public=False, enabled=False)
+def test_explicit_public_permission_null_counts_as_update() -> None:
+    req = SkillPatchRequest(public_permission=None, enabled=False)
 
-    assert req.model_fields_set == {"is_public", "enabled"}
+    assert req.model_fields_set == {"public_permission", "enabled"}
     assert req.has_db_field_update is True
-    assert req.is_public is False
+    assert req.public_permission is None
     assert req.enabled is False
 
 
@@ -65,8 +63,6 @@ def test_detail_fields_track_intent() -> None:
         "name",
         "description",
         "instructions_markdown",
-        "is_public",
-        "public_permission",
         "enabled",
     ],
 )
@@ -74,6 +70,14 @@ def test_explicit_null_rejected(field: str) -> None:
     """Sending field=null (not omitting it) should raise ValidationError."""
     with pytest.raises(ValidationError, match=f"{field} cannot be null"):
         SkillPatchRequest.model_validate({field: None})
+
+
+def test_public_permission_null_allowed() -> None:
+    req = SkillPatchRequest.model_validate({"public_permission": None})
+
+    assert req.public_permission is None
+    assert req.model_fields_set == {"public_permission"}
+    assert req.has_db_field_update is True
 
 
 @pytest.mark.parametrize("field", ["name", "description", "instructions_markdown"])
