+++ web/src/lib/skills/__fixtures__/picker.ts
@@ -13,7 +13,6 @@ export function builtinFixture(over: Partial<BuiltinSkill> = {}): BuiltinSkill {
     description: "Build PowerPoint decks.",
     is_available: true,
     unavailable_reason: null,
-    is_public: null,
     is_personal: false,
     enabled: null,
     author_user_id: null,
@@ -39,7 +38,6 @@ export function customFixture(over: Partial<CustomSkill> = {}): CustomSkill {
     description: "Draft a structured report from notes.",
     is_available: null,
     unavailable_reason: null,
-    is_public: true,
     is_personal: false,
     enabled: true,
     author_user_id: null,
